+++ src/app/read/experimental/[identifier]/page.tsx
@@ -2,7 +2,7 @@
 
 import { use, useEffect, useState } from "react";
 import { ExperimentalWebPubStatefulReader } from "@/components/WebPub";
-import { StatefulLoader } from "@/components/StatefulLoader";
+import { StatefulLoader } from "@/components/Misc/StatefulLoader";
 import { usePublication } from "@/hooks/usePublication";
 import { useAppSelector } from "@/lib/hooks";
 import { verifyManifestUrl } from "@/app/api/verify-manifest/verifyDomain";
