+++ src/components/index.ts
@@ -5,8 +5,6 @@ export * from "./Docking";
 export * from "./Plugins";
 export * from "./Settings";
 export * from "./Sheets";
-export * from "./StatefulLoader";
-export * from "./PublicationGrid";
 export * from "./StatefulPreferencesProvider";
 
 // export * from "../StatefulReaderArrowButton";
