+++ tsup.config.ts
@@ -10,12 +10,13 @@ export default defineConfig({
     "src/core/Components/index.ts",
     "src/core/Helpers/index.ts",
     "src/core/Hooks/index.ts", 
+    "src/components/Misc/index.ts",
     "src/components/Epub/index.ts",
+    "src/components/WebPub/index.ts",
     "src/i18n/index.ts",
     "src/lib/index.ts",
     "src/preferences/index.ts",
-    "src/next-lib/index.ts",
-    "src/components/WebPub/index.ts"
+    "src/next-lib/index.ts"
   ],
   loader: {
     ".css": "local-css"
